AnimateGAN/requirements.txt
altair==4.2.0
asn1crypto==0.24.0
attrs==22.1.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
blinker==1.5
cachetools==5.2.0
certifi==2018.1.18
cffi==1.15.1
chardet==3.0.4
click==8.1.3
commonmark==0.9.1
cryptography==38.0.3
decorator==5.1.1
dfn==1.0.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.4
gitdb==4.0.10
GitPython==3.1.29
idna==2.6
importlib-metadata==5.1.0
importlib-resources==5.10.0
iniconfig==1.1.1
Jinja2==3.1.2
jsonschema==4.17.1
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==2.1.1
nasops==0.0.1
numpy==1.23.5
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
paramiko==2.12.0
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
prettytable==0.7.2
protobuf==3.20.3
py==1.11.0
pyaml==17.12.1
pyarrow==10.0.1
pyasn1==0.4.2
pycparser==2.21
pycrypto==2.6.1
pydeck==0.8.0
Pygments==2.13.0
PyGObject==3.26.1
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.2
pyserial==3.5
pytest==7.2.0
pytest-assume==2.4.3
pytest-dependency==0.5.1
pytest-html==3.2.0
pytest-metadata==2.0.4
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.25
PyYAML==3.12
regex==2022.10.31
requests==2.18.4
rich==12.6.0
SecretStorage==2.3.1
semver==2.13.0
six==1.11.0
smmap==5.0.0
streamlit==1.15.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
Unidecode==1.0.22
urllib3==1.22
validators==0.20.0
watchdog==2.1.9
wget==3.2
zipp==3.11.0
